Add elemental database back to public API
