+++ exspy/__init__.py
@@ -24,7 +24,7 @@ from . import data
 from . import models
 from . import signals
 from . import utils
-from ._misc import material
+from . import material
 from ._defaults_parser import preferences
